function_app/requirements.txt
azure-functions==1.23.0
cuid==0.4
MarkupSafe==3.0.2
psycopg2==2.9.10
PyJWT==2.10.1
Werkzeug==3.1.3
